Merge in RFRM/pay from MGMPROCESS-371-refactor to develop

* commit '4aeacb3de9c30f6f21aaf6b24dc3526eac7c822f':
  Рефакторинг обработки ответа от 1503, изменение на List<Map<>>.

+++ src/main/java/ru/vtb/msa/rfrm/functions/FunctionPDImpl.java
@@ -50,7 +50,7 @@ public class FunctionPDImpl implements FunctionPD {
     @SneakyThrows
     private void handleMdmIdList(List<EntPaymentTask> entPaymentTaskList) {
 
-        List<Integer> setRewardIdList = entPaymentTaskList.stream().map(EntPaymentTask::getRewardId).distinct().collect(Collectors.toList());
+        List<Long> setRewardIdList = entPaymentTaskList.stream().map(EntPaymentTask::getRewardId).distinct().collect(Collectors.toList());
 
         //Установить для задачи blocked=1 и blocked_at=now()
         actionEntPaymentTaskRepo.updateBlockByRewardIdEqualOne(setRewardIdList);
